# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 电气车间备件管理系统
  version: 1.0.0
paths:
...
  /api/v1/inventory/balances:
    get:
      tags:
      - 库存
      summary: Balances
      operationId: balances_api_v1_inventory_balances_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      - name: min_qty
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Min Qty
      - name: max_qty
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Max Qty
      - name: low_stock
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Low Stock
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_InventoryBalanceRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/inventory/low-stock:
    get:
      tags:
      - 库存
      summary: Low Stock
      operationId: low_stock_api_v1_inventory_low_stock_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: keyword
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Keyword
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_InventoryBalanceRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/inventory/inbounds:
    post:
      tags:
      - 库存
      summary: Inbound
      operationId: inbound_api_v1_inventory_inbounds_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOperationRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - HTTPBearer: []
  /api/v1/inventory/outbounds:
    post:
      tags:
      - 库存
      summary: Outbound
      operationId: outbound_api_v1_inventory_outbounds_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOperationRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - HTTPBearer: []
  /api/v1/inventory/operations:
    get:
      tags:
      - 库存
      summary: Operations
      operationId: operations_api_v1_inventory_operations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 20
          title: Page Size
      - name: operation_no
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Operation No
      - name: operation_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/OperationType'
          - type: 'null'
          title: Operation Type
      - name: material_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Material Name
      - name: operator_name
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Operator Name
      - name: purchase_request_no
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Purchase Request No
      - name: source_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/SourceType'
          - type: 'null'
          title: Source Type
      - name: start_at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Start At
      - name: end_at
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: End At
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_StockOperationRead_'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/inventory/operations/{operation_id}:
    get:
      tags:
      - 库存
      summary: Operation Detail
      operationId: operation_detail_api_v1_inventory_operations__operation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: integer
          title: Operation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOperationRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - 库存
      summary: Edit Operation
      operationId: edit_operation_api_v1_inventory_operations__operation_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: integer
          title: Operation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOperationRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/inventory/operations/{operation_id}/reverse:
    post:
      tags:
      - 库存
      summary: Reverse Operation
      operationId: reverse_operation_api_v1_inventory_operations__operation_id__reverse_post
      security:
      - HTTPBearer: []
      parameters:
      - name: operation_id
        in: path
        required: true
        schema:
          type: integer
          title: Operation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReverseOperationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockOperationRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/inventory/low-stock/{material_id}/create-replenishment-draft:
    post:
      tags:
      - 库存
      summary: Replenish
      operationId: replenish_api_v1_inventory_low_stock__material_id__create_replenishment_draft_post
      security:
      - HTTPBearer: []
      parameters:
      - name: material_id
        in: path
        required: true
        schema:
          type: integer
          title: Material Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplenishmentDraftRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/dashboard/summary:
    get:
      tags:
      - 库存
      summary: Summary
      operationId: summary_api_v1_dashboard_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummaryRead'
        '400':
          description: 业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: 未认证或凭证无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 版本、状态或业务数据冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: 请求参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          additionalProperties: true
          type: object
          title: Details
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - code
      - message
      - request_id
      title: ApiError
...
    DashboardSummaryRead:
      properties:
        stock_material_count:
          type: integer
          title: Stock Material Count
        low_stock_count:
          type: integer
          title: Low Stock Count
        pending_purchase_count:
          type: integer
          title: Pending Purchase Count
        on_order_count:
          type: integer
          title: On Order Count
        uncoded_purchase_material_count:
          type: integer
          title: Uncoded Purchase Material Count
        pending_purchase_request_count:
          type: integer
          title: Pending Purchase Request Count
        partially_received_count:
          type: integer
          title: Partially Received Count
      type: object
      required:
      - stock_material_count
      - low_stock_count
      - pending_purchase_count
      - on_order_count
      - uncoded_purchase_material_count
      - pending_purchase_request_count
      - partially_received_count
      title: DashboardSummaryRead
...
    InventoryBalanceRead:
      properties:
        stock_material_id:
          type: integer
          title: Stock Material Id
        name:
          type: string
          title: Name
        model_spec:
          type: string
          title: Model Spec
        unit_name:
          type: string
          title: Unit Name
        decimal_places:
          type: integer
          title: Decimal Places
        current_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Current Qty
        minimum_qty:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Minimum Qty
        target_qty:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Target Qty
        on_order_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: On Order Qty
        is_low_stock:
          type: boolean
          title: Is Low Stock
        warning_state:
          anyOf:
          - type: string
            enum:
            - PENDING_PURCHASE
            - ON_ORDER
          - type: 'null'
          title: Warning State
        suggested_purchase_qty:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Suggested Purchase Qty
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - stock_material_id
      - name
      - model_spec
      - unit_name
      - decimal_places
      - current_qty
      - on_order_qty
      - is_low_stock
      - suggested_purchase_qty
      - updated_at
      title: InventoryBalanceRead
...
    OperationCreate:
      properties:
        client_request_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Client Request Id
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        source_type:
          $ref: '#/components/schemas/SourceType'
        business_reason:
          type: string
          maxLength: 500
          minLength: 1
          title: Business Reason
        receiver_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Receiver Name
        project_subitem_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Subitem Id
        lines:
          items:
            $ref: '#/components/schemas/OperationLineWrite'
          type: array
          minItems: 1
          title: Lines
      additionalProperties: false
      type: object
      required:
      - client_request_id
      - occurred_at
      - source_type
      - business_reason
      - lines
      title: OperationCreate
...
    OperationType:
      type: string
      enum:
      - INBOUND
      - OUTBOUND
      title: OperationType
    OperationUpdate:
      properties:
        version:
          type: integer
          title: Version
        operation_type:
          $ref: '#/components/schemas/OperationType'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        source_type:
          $ref: '#/components/schemas/SourceType'
        business_reason:
          type: string
          maxLength: 500
          minLength: 1
          title: Business Reason
        receiver_name:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Receiver Name
        project_subitem_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Subitem Id
        lines:
          items:
            $ref: '#/components/schemas/OperationLineWrite'
          type: array
          minItems: 1
          title: Lines
      additionalProperties: false
      type: object
      required:
      - version
      - operation_type
      - occurred_at
      - source_type
      - business_reason
      - lines
      title: OperationUpdate
    Page_InventoryBalanceRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/InventoryBalanceRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[InventoryBalanceRead]
...
    Page_StockOperationRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/StockOperationRead'
          type: array
          title: Items
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        total:
          type: integer
          title: Total
      type: object
      required:
      - items
      - page
      - page_size
      - total
      title: Page[StockOperationRead]
...
    ReplenishmentDraftRead:
      properties:
        next:
          type: string
          const: purchase_material
          title: Next
        resource_id:
          type: integer
          title: Resource Id
      type: object
      required:
      - next
      - resource_id
      title: ReplenishmentDraftRead
...
    ReverseOperationRequest:
      properties:
        client_request_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Client Request Id
        reason:
          type: string
          maxLength: 500
          minLength: 1
          title: Reason
      additionalProperties: false
      type: object
      required:
      - client_request_id
      - reason
      title: ReverseOperationRequest
...
    SourceType:
      type: string
      enum:
      - MANUAL
      - PURCHASE_RECEIPT
      - REVERSAL
      - INITIALIZATION
      title: SourceType
...
    StockOperationRead:
      properties:
        id:
          type: integer
          title: Id
        operation_no:
          type: string
          title: Operation No
        operation_type:
          $ref: '#/components/schemas/OperationType'
        occurred_at:
          type: string
          format: date-time
          title: Occurred At
        operator_name:
          type: string
          title: Operator Name
        business_reason:
          type: string
          title: Business Reason
        receiver_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Receiver Name
        project_subitem_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Subitem Id
        source_type:
          $ref: '#/components/schemas/SourceType'
        reversal_of_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reversal Of Id
        purchase_request_no:
          anyOf:
          - type: string
          - type: 'null'
          title: Purchase Request No
        client_request_id:
          type: string
          title: Client Request Id
        lines:
          items:
            $ref: '#/components/schemas/StockOperationLineRead'
          type: array
          title: Lines
        created_at:
          type: string
          format: date-time
          title: Created At
        version:
          type: integer
          title: Version
      type: object
      required:
      - id
      - operation_no
      - operation_type
      - occurred_at
      - operator_name
      - business_reason
      - source_type
      - client_request_id
      - lines
      - created_at
      - version
      title: StockOperationRead
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer